+++ backend/src/main/java/com/geekbrains/practice/entry/ServerInitializer.java
@@ -1,15 +1,17 @@
 package com.geekbrains.practice.entry;
 
-import com.geekbrains.practice.listeners.UserCheckerListener;
+import com.geekbrains.practice.services.UserCheckerAuthService;
 
 import java.io.IOException;
 import java.net.ServerSocket;
+import java.net.Socket;
 
 public class ServerInitializer {
   public static void main(String[] args) {
     try {
       ServerSocket serverSocket = new ServerSocket(8189);
-      new UserCheckerListener(serverSocket.accept()).startListening();
+      Socket socket = serverSocket.accept();
+      new UserCheckerAuthService(socket).startListening();
     } catch (IOException e) {
       e.printStackTrace();
     }
