--- backend/src/main/java/com/geekbrains/practice/listeners/IListener.java
@@ -1,5 +1,0 @@
-package com.geekbrains.practice.listeners;
-
-public interface IListener {
-  void startListening();
-}
